requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
asgiref==3.6.0
    # via django
django==4.2
    # via -r requirements.in
sqlparse==0.4.4
    # via django

djangorestframework~=3.14.0